requirements/base.txt
Django==4.1.5
django-environ==0.8.1
djangorestframework==3.14.0
django-cors-headers==3.13.0
django-filter==21.1
django-autoslug==1.9.8
django-countries==7.2.1
django-phonenumber-field==5.2.0
phonenumbers==8.12.33
drf-yasg==1.21.5


environ==1.0


#Pillow==9.0.1


argon2-cffi==21.3.0


pytz==2021.3